refactor: Moving PageLoading up to the top
This way it can be used on both the courseware and outline pages.
